+++ open_spiel/games/catch.cc
@@ -160,7 +160,7 @@ std::string CatchState::ObservationString(Player player) const {
 }
 
 void CatchState::ObservationTensor(Player player,
-                                   std::vector<double>* values) const {
+                                   std::vector<float>* values) const {
   SPIEL_CHECK_GE(player, 0);
   SPIEL_CHECK_LT(player, num_players_);
 
@@ -172,7 +172,7 @@ void CatchState::ObservationTensor(Player player,
 }
 
 void CatchState::InformationStateTensor(Player player,
-                                        std::vector<double>* values) const {
+                                        std::vector<float>* values) const {
   SPIEL_CHECK_GE(player, 0);
   SPIEL_CHECK_LT(player, num_players_);
