+++ open_spiel/games/cliff_walking.h
@@ -88,9 +88,9 @@ class CliffWalkingState : public State {
   std::string InformationStateString(Player player) const override;
   std::string ObservationString(Player player) const override;
   void ObservationTensor(Player player,
-                         std::vector<double>* values) const override;
+                         std::vector<float>* values) const override;
   void InformationStateTensor(Player player,
-                              std::vector<double>* values) const override;
+                              std::vector<float>* values) const override;
   std::unique_ptr<State> Clone() const override;
   void UndoAction(Player player, Action move) override;
   std::vector<Action> LegalActions() const override;
