+++ open_spiel/games/cursor_go.cc
@@ -110,7 +110,7 @@ std::string CursorGoState::ObservationString(Player player) const {
 }
 
 void CursorGoState::ObservationTensor(Player player,
-                                      std::vector<double>* values) const {
+                                      std::vector<float>* values) const {
   SPIEL_CHECK_GE(player, 0);
   SPIEL_CHECK_LT(player, num_players_);
