+++ open_spiel/games/go.cc
@@ -104,7 +104,7 @@ std::string GoState::ObservationString(int player) const {
   return ToString();
 }
 
-void GoState::ObservationTensor(int player, std::vector<double>* values) const {
+void GoState::ObservationTensor(int player, std::vector<float>* values) const {
   SPIEL_CHECK_GE(player, 0);
   SPIEL_CHECK_LT(player, num_players_);
