+++ open_spiel/games/goofspiel.cc
@@ -486,7 +486,7 @@ void GoofspielState::NextPlayer(int* count, Player* player) const {
 }
 
 void GoofspielState::InformationStateTensor(Player player,
-                                            std::vector<double>* values) const {
+                                            std::vector<float>* values) const {
   SPIEL_CHECK_GE(player, 0);
   SPIEL_CHECK_LT(player, num_players_);
 
@@ -566,7 +566,7 @@ void GoofspielState::InformationStateTensor(Player player,
 }
 
 void GoofspielState::ObservationTensor(Player player,
-                                       std::vector<double>* values) const {
+                                       std::vector<float>* values) const {
   SPIEL_CHECK_GE(player, 0);
   SPIEL_CHECK_LT(player, num_players_);
