+++ open_spiel/games/hanabi.h
@@ -79,7 +79,7 @@ class OpenSpielHanabiState : public State {
   // The observation by default includes knowledge inferred from past hints.
   std::string ObservationString(Player player) const override;
   void ObservationTensor(Player player,
-                         std::vector<double>* values) const override;
+                         std::vector<float>* values) const override;
 
   std::unique_ptr<State> Clone() const override;
   ActionsAndProbs ChanceOutcomes() const override;
