+++ open_spiel/games/negotiation.h
@@ -84,7 +84,7 @@ class NegotiationState : public State {
   std::vector<double> Returns() const override;
   std::string ObservationString(Player player) const override;
   void ObservationTensor(Player player,
-                         std::vector<double>* values) const override;
+                         std::vector<float>* values) const override;
 
   std::unique_ptr<State> Clone() const override;
   std::vector<Action> LegalActions() const override;
