+++ open_spiel/games/pentago.cc
@@ -266,7 +266,7 @@ int PlayerRelative(PentagoPlayer state, Player current) {
 }
 
 void PentagoState::ObservationTensor(Player player,
-                                     std::vector<double>* values) const {
+                                     std::vector<float>* values) const {
   SPIEL_CHECK_GE(player, 0);
   SPIEL_CHECK_LT(player, num_players_);
